backend/pyproject.toml
[project]
name = "oozo-backend"
version = "1.0.0"
description = "Backend for OOZO document analysis system"
authors = [
    {name = "OOZO Team", email = "team@oozo.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "faiss-cpu>=1.7.4",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "sentence-transformers>=2.2.2",
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.1",
    "pdfminer.six>=20221105",
    "python-multipart>=0.0.6",
    "openai>=1.57.0",
    "nltk>=3.8.1",
    "spacy>=3.7.2",
    "textstat>=0.7.3",
    "pydantic>=2.10.3",
    "email-validator>=2.1.0",
    "httpx>=0.27.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dateutil>=2.9.0.post0",
    "prometheus_client>=0.19.0",
    "loguru>=0.7.2",
    "pydantic-settings>=2.1.0",
    "chardet>=5.2.0",
    "openpyxl>=3.1.2",
    "redis>=5.0.1",
    "factory-boy>=3.3.0",
    "faker>=20.1.0"
]
requires-python = ">=3.8.1,<4.0"

[tool.hatch.build.targets.wheel]
packages = ["app"]
